refator input organization
