+++ src/components/event_maker/pages/DynamicComponentConfiguration.jsx
@@ -3,9 +3,11 @@ import PropTypes from "prop-types";
 import { FaChevronDown, FaChevronUp } from "react-icons/fa";
 import { useComponents } from "../../../contexts/ComponentsContext";
 import { SortableItemLayout } from "./SortableItemLayout";
-import { TextInput, CheckboxInput, SelectInput } from "./CommonInputs";
-import { NumberInput } from "../../forms/NumberInput.jsx";
-import { MediaInput } from "../../../components/forms/MediaInput";
+import { TextInput } from "./inputs/TextInput.jsx";
+import { CheckboxInput } from "./inputs/CheckboxInput.jsx";
+import { SelectInput } from "./inputs/SelectInput.jsx";
+import { NumberInput } from "./inputs/NumberInput.jsx";
+import { MediaInput } from "./inputs/MediaInput.jsx";
 import { ColorSelector } from "../ColorSelector.jsx";
 import { ColorThemeLayout } from "../ColorThemeLayout.jsx";
 import * as Components from "./components";
